SPIGOT-4608: Improve quality of MapMeta APIs
By: md_5 <git@md-5.net>

+++ paper-api/src/main/java/org/bukkit/Bukkit.java
@@ -521,7 +521,7 @@ public final class Bukkit {
      * @deprecated Magic value
      */
     @Deprecated
-    public static MapView getMap(short id) {
+    public static MapView getMap(int id) {
         return server.getMap(id);
     }
